requirements.txt
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
streamlit==1.25.0
nltk==3.8.1
wordcloud==1.9.2
scipy==1.11.1
joblib==1.3.1
tqdm==4.65.0
textblob==0.17.1